# open_api_docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/quiz-questions/{quizQuestion}/toggle-approval:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: toggleApproval
      parameters:
      - name: quizQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
  /api/quiz-questions/{quizQuestion}/suggest-fine-tuning:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: suggestQuestionForFineTuning
      parameters:
      - name: quizQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSuggestionCreationParams"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/quiz-questions/{quizQuestion}/regenerate:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: regenerate
      parameters:
      - name: quizQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestion"
  /api/quiz-questions/{quizQuestion}/contest:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: contest
      parameters:
      - name: quizQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionContestResult"
  /api/quiz-questions/{quizQuestion}/answer:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: answerQuiz
      parameters:
      - name: quizQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerDTO"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AnsweredQuestion"
  /api/quiz-questions/{note}/refine-question:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: refineQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionAndAnswer"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
  /api/quiz-questions/{note}/note-questions:
    get:
      tags:
      - rest-quiz-question-controller
      operationId: getAllQuestionByNote
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestionAndAnswer"
    post:
      tags:
      - rest-quiz-question-controller
      operationId: addQuestionManually
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionAndAnswer"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
    delete:
      tags:
      - rest-quiz-question-controller
      operationId: deleteQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionAndAnswer"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
    patch:
      tags:
      - rest-quiz-question-controller
      operationId: updateQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionAndAnswer"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
  /api/quiz-questions/generate-question:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: generateQuestion
      parameters:
      - name: note
        in: query
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionInNotebook"
  /api/quiz-questions/generate-question-without-save:
    post:
      tags:
      - rest-quiz-question-controller
      operationId: generateAIQuestionWithoutSave
      parameters:
      - name: note
        in: query
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuizQuestionAndAnswer"
...
components:
  schemas:
...
    MCQWithAnswer:
      required:
      - correctChoiceIndex
      - multipleChoicesQuestion
      type: object
      properties:
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctChoiceIndex:
          type: integer
          format: int32
    MultipleChoicesQuestion:
      required:
      - choices
      type: object
      properties:
        stem:
          type: string
        choices:
          type: array
          items:
            type: string
...
    ImageWithMask:
      type: object
      properties:
        noteImage:
          type: string
        imageMask:
          type: string
    QuizQuestion:
      required:
      - id
      - multipleChoicesQuestion
      type: object
      properties:
        id:
          type: integer
          format: int32
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        checkSpell:
          type: boolean
        imageWithMask:
          $ref: "#/components/schemas/ImageWithMask"
    QuizQuestionAndAnswer:
      required:
      - id
      - quizQuestion
      type: object
      properties:
        id:
          type: integer
          format: int32
        quizQuestion:
          $ref: "#/components/schemas/QuizQuestion"
        correctAnswerIndex:
          type: integer
          format: int32
        approved:
          type: boolean
...
    Note:
      required:
      - createdAt
      - id
      - noteTopic
      - updatedAt
      type: object
      properties:
        noteTopic:
          $ref: "#/components/schemas/NoteTopic"
        details:
          type: string
        parentId:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          readOnly: true
        wikidataId:
          type: string
    NoteTopic:
      required:
      - id
      - topicConstructor
      type: object
      properties:
        id:
          type: integer
          format: int32
        topicConstructor:
          type: string
        shortDetails:
          type: string
        linkType:
          type: string
          enum:
          - no link
          - related to
          - a specialization of
          - an application of
          - an instance of
          - a part of
          - tagged by
          - an attribute of
          - the opposite of
          - author of
          - using
          - an example of
          - before
          - similar to
          - confused with
        targetNoteTopic:
          $ref: "#/components/schemas/NoteTopic"
        parentNoteTopic:
          $ref: "#/components/schemas/NoteTopic"
    Notebook:
      required:
      - headNote
      - id
      - notebookSettings
      type: object
      properties:
        id:
          type: integer
          format: int32
        headNote:
          $ref: "#/components/schemas/Note"
        notebookSettings:
          $ref: "#/components/schemas/NotebookSettings"
        creatorId:
          type: string
    NotebookSettings:
      type: object
      properties:
        skipReviewEntirely:
          type: boolean
        numberOfQuestionsInAssessment:
          type: integer
          format: int32
...
    ReviewPoint:
      required:
      - id
      - nextReviewAt
      - note
      type: object
      properties:
        id:
          type: integer
          format: int32
        note:
          $ref: "#/components/schemas/Note"
        lastReviewedAt:
          type: string
          format: date-time
        nextReviewAt:
          type: string
          format: date-time
        initialReviewedAt:
          type: string
          format: date-time
        repetitionCount:
          type: integer
          format: int32
        forgettingCurveIndex:
          type: integer
          format: int32
        removedFromReview:
          type: boolean
...
    QuestionSuggestionCreationParams:
      type: object
      properties:
        comment:
          type: string
        isPositiveFeedback:
          type: boolean
    SuggestedQuestionForFineTuning:
      required:
      - id
      - preservedQuestion
      type: object
      properties:
        id:
          type: integer
          format: int32
        comment:
          type: string
        preservedQuestion:
          $ref: "#/components/schemas/MCQWithAnswer"
        preservedNoteContent:
          type: string
        realCorrectAnswers:
          type: string
        createdAt:
          type: string
          format: date-time
        positiveFeedback:
          type: boolean
    QuizQuestionContestResult:
      type: object
      properties:
        reason:
          type: string
        rejected:
          type: boolean
    AnswerDTO:
      type: object
      properties:
        spellingAnswer:
          type: string
        choiceIndex:
          type: integer
          format: int32
    AnsweredQuestion:
      type: object
      properties:
        answerId:
          type: integer
          format: int32
        correct:
          type: boolean
        correctChoiceIndex:
          type: integer
          format: int32
        choiceIndex:
          type: integer
          format: int32
        answerDisplay:
          type: string
        reviewPoint:
          $ref: "#/components/schemas/ReviewPoint"
        quizQuestionAndAnswer:
          $ref: "#/components/schemas/QuizQuestionAndAnswer"
    QuizQuestionInNotebook:
      required:
      - notebook
      - quizQuestion
      type: object
      properties:
        notebook:
          $ref: "#/components/schemas/Notebook"
        quizQuestion:
          $ref: "#/components/schemas/QuizQuestion"
...